+++ main-window.cpp
@@ -255,13 +255,13 @@ MainWindow::MainWindow(QWidget* const parent)
 
 						for (auto &piece : mappings->frames[frame_index].pieces)
 						{
-							const unsigned int base_mapped_tile = dplc_frame.getMappedTile(piece.tile_index);
+							const int base_mapped_tile = dplc_frame.getMappedTile(piece.tile_index);
 
-							for (unsigned int i = 0; i < piece.width * piece.height; ++i)
+							for (int i = 0; i < piece.width * piece.height; ++i)
 							{
-								const unsigned int mapped_tile = dplc_frame.getMappedTile(piece.tile_index + i);
+								const int mapped_tile = dplc_frame.getMappedTile(piece.tile_index + i);
 
-								if (mapped_tile == static_cast<unsigned int>(-1) || mapped_tile != base_mapped_tile + i)
+								if (mapped_tile == -1 || mapped_tile != base_mapped_tile + i)
 									return;
 							}
